first step of restructering
